requirements.txt
streamlit==1.24.0
numpy==1.23.0
pillow==9.5.0
matplotlib==3.7.1
tensorflow==2.12.0
keras==2.12.0